+++ src/collective/solr/interfaces.py
@@ -147,10 +147,12 @@ class ISolrSchema(Interface):
 
     force_simple_search = Bool(
         title=_("label_force_simple_search",
-                default=u"Force the use of the Pattern for FIXME"),
+                default=u"Force simple search pattern"),
         description=_(
             "help_force_simple_search",
-            default=u"The search will always use the Pattern for simple search.",
+            default=u"If set all queries in SearchableText will use the "
+            u"\"Pattern for simple search queries\". This will remove all special "
+            u"solr characters and operators from the query text.",
         ),
     )
 
@@ -159,10 +161,9 @@ class ISolrSchema(Interface):
                 default=u"Allow complex search"),
         description=_(
             "help_allow_complex_search",
-            default=u"If \"Force the use of the Pattern for FIXME\" is set "
-            u"allow to pass throuh a complex solr pattern for SearchableText "
-            u"by either prefixing SearchableText with \"solr:\" or passing "
-            u"solr_complex_search=True in the catalog query.",
+            default=u"Allow a complex search in SearchableText by either prefixing the "
+            u"search text with \"solr:\" or passing \"solr_complex_search\" as a field "
+            u"in the catalog query.",
         ),
     )